add proper support for forks
it is now possible to trigger jenkins also on pull requests which were
created from a fork.

two new variables were introduced for this:
$PRSOURCEPROJECT:    the PROJECT where the source repository resides
$PRSOURCEREPOSITORY: the REPOSITORY where the source branch resides

the existing $PROJECT and $REPOSITORY will continue to point to the
destination, even though it might have been more logical to flip the two
(instead introducing $PRTARGET[PROJECT|REPOSITORY]) so that all jobs can
just use $PROJECT and $REPOSITORY and not care whether it is a pull
request or not. however, this would be a breaking change for existing
consumers which use $PROJECT and $REPOSITORY in a fork-setup on pull
requests (though they currently can't actually build the project since
they lack the necessary information).

note that the new code follows the inconsistent coding style of the
existing code to be aligned at least with the few lines surrounding it.
this also means that in the same commit different coding styles were
mixed :(

this closes KyleLNicholls/parameterized-builds#194